% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.scripting.R
\name{cmd.iqtree}
\alias{cmd.iqtree}
\title{Produce a single iqtree shell command.}
\usage{
cmd.iqtree(
  infile.fasta,
  outfile = paste(infile.fasta, ".phy", sep = ""),
  pr = PR,
  pr.args = "-m GTRCAT --HKY85 -p 42"
)
}
\value{
Character string
}
\description{
Produce a single iqtree shell command.
}
