% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.clustering.R
\name{hivc.clu.brdist.stats}
\alias{hivc.clu.brdist.stats}
\title{Compute a statistic of patristic distances for each subtree in the tree}
\usage{
hivc.clu.brdist.stats(tree, distmat = NULL, eval.dist.btw = "leaf",
  stat.fun = max)
}
\description{
Compute a statistic of patristic distances for each subtree in the tree
}

