% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.bam.get.length.and.pos.of.mergedreads}
\alias{phsc.bam.get.length.and.pos.of.mergedreads}
\title{Calculate position and length of merged reads}
\usage{
phsc.bam.get.length.and.pos.of.mergedreads(bam.file.name, error.strict = TRUE)
}
\arguments{
\item{bam.file.name}{full path name to bam file.}
}
\value{
data.table with columns QNAME (template query ID), POS (leftmost position of read), LEN (length of read)
}
\description{
This function calculates the position and length of the two sequenced segments from a single RNA template, potentially after merging when both segments overlap.
}
