% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.scripting.R
\name{phsc.cmd.bam.calculate.read.distribution}
\alias{phsc.cmd.bam.calculate.read.distribution}
\title{Generate bash commands to calculate read distribution in bam file}
\usage{
phsc.cmd.bam.calculate.read.distribution(pty.runs, pty.args)
}
\arguments{
\item{pty.runs}{data.table with columns SAMPLE_ID and PTY_RUN}

\item{pty.args}{List of input variables containing the fields "prog.bam.distr.calculator", "data.dir", "out.dir", "work.dir"}
}
\value{
character string of bash commands.
}
\description{
This function generates bash commands to calculate read distribution in bam file
}
