% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.cmd.phyloscanner.multi}
\alias{phsc.cmd.phyloscanner.multi}
\title{Generate bash commands for a multiple phyloscanner runs}
\usage{
phsc.cmd.phyloscanner.multi(pty.runs, pty.args)
}
\description{
This function generates bash commands for multiple phyloscanner runs, that can be called via 'system' in R, or written to file to run on a UNIX system.
}

