% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.cmd.phyloscanner.one}
\alias{phsc.cmd.phyloscanner.one}
\title{Generate bash command for a single phyloscanner run}
\usage{
phsc.cmd.phyloscanner.one(file.bam, file.ref, pty.args)
}
\arguments{
\item{file.bam}{File name of the file that contains the list of bam files.}

\item{file.ref}{File name of the file that contains the list of reference files.}

\item{pty.args}{List of phyloscanner input variables.}
}
\description{
This function generates bash commands for a single phyloscanner run, that can be called via 'system' in R, or written to file to run on a UNIX system.
}

