% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.cmd.phyloscanner.one.resume.combinewindows}
\alias{phsc.cmd.phyloscanner.one.resume.combinewindows}
\title{Generate bash command to combine single window resumes of a phyloscanner run}
\usage{
phsc.cmd.phyloscanner.one.resume.combinewindows(prefix.infiles, pty.args)
}
\arguments{
\item{prefix.infiles}{File name that points phyloscanner output.}

\item{pty.args}{List of phyloscanner input variables. See examples.}
}
\value{
Character string of phyloscanner commands.
}
\description{
This function generates bash commands to resume a single phyloscanner run, from the point where all read alignments and read phylogenies were created. The bash script can be called via 'system' in R, or written to file to run on a UNIX system.
}
\keyword{internal}
