% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.cmd.process.phyloscanner.output.in.directory}
\alias{phsc.cmd.process.phyloscanner.output.in.directory}
\title{Generate bash commands to process phyloscanner output}
\usage{
phsc.cmd.process.phyloscanner.output.in.directory(pty.args, run.id, tmp.dir,
  file.bam)
}
\arguments{
\item{pty.args}{phyloscanner input arguments that were used to obtain phyloscanner output.}

\item{run.id}{Character string that identifies phyloscanner output.}

\item{tmp.dir}{Directory with phyloscanner output.}

\item{file.bam}{File name of the file that contains the list of bam files.}
}
\value{
data.table with one column CMD that contains the bash commands.
}
\description{
This function generates bash commands that combine the various Rscripts in the phyloscanner toolkit
}

