% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.combine.phyloscanner.output}
\alias{phsc.combine.phyloscanner.output}
\title{Combine output from multiple phyloscanner runs}
\usage{
phsc.combine.phyloscanner.output(in.dir, save.file = NA,
  postfix.trees = "trees.rda",
  postfix.trmwindowstats = "trmStatsPerWindow.rda",
  regex.ind = "^[0-9]+_[0-9]_[0-9]+", trmw.min.reads = 1,
  trmw.min.tips = 1, trmw.close.brl = Inf, trmw.distant.brl = Inf,
  norm.file.name = NA)
}
\arguments{
\item{in.dir}{Full directory path to output of phyloscanner runs}

\item{save.file}{If not NA, the combined output is saved to file.}

\item{postfix.trees}{Pattern at end of file names to identify short read tree output generated by the phyloscanner tools. By default 'trees.rda'.}

\item{postfix.trmwindowstats}{Pattern at end of file names to identify transmission statistics output per window, generated by the phyloscanner tools. By default 'trmStatsPerWindow.rda'.}

\item{regex.ind}{Regular expression that identifies the ID of query individuals}

\item{trmw.min.reads}{Minimum number of reads per individuals in order to make a transmission assignment that involves this individual}
}
\value{
list of three R objects 'phs', 'dtrees', 'dtrms'. See description.
}
\description{
This function generates three R objects. 'phs' is a list of all short read trees in ape format.
	  'dtrees' is a data.table that provides info on each short read tree. Columns are 'PTY_RUN' (phyloscanner run id), 'W_FROM' (start of window), 'W_TO' (end of window), 'IDX' (index of tree in phs).
	  'dtrms' is a data.table that provides info on transmission assignments. Columns are 'PTY_RUN' (phyloscanner run id), 'ID1' (identifier of first individual), 'ID2' (identifier of second individual), 'TYPE' (window assignment), 
  'WIN_OF_TYPE' (number of windows with that assignment), 'WIN_TOTAL' (Number of windows where reads from both individuals are present), 'PAIR_ID' (unique ID of each combination of PTY_RUN,ID1, ID2)
}
