% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.define.stage1.analyses}
\alias{phsc.define.stage1.analyses}
\title{Determine stage 1 phyloscanner runs}
\usage{
phsc.define.stage1.analyses(x, batch.size = 50)
}
\arguments{
\item{x}{Character vector of individual identifiers.}

\item{batch.size}{Batch size. Default is 50.}
}
\value{
data.table with rows 'IND' (individual identifiers), 'PTY_RUN' group for phyloscanner analysis, and 'BATCH' batch of individuals (not used further, but there should be two batches of individuals in each phyloscanner analysis).
}
\description{
This function groups individuals for phyloscanner analyses, so that phylogenetic linkage between every pair of individuals is assessed at least once.
  Specifically, individuals are grouped into batches of specified size, and then, all possible pairs of batches are formed. 
  Each of these pairs of batches defines a group of individuals between whom phylogenetic linkages are assessed in one phyloscanner run.
  The number of individuals in each group is twice the batch size.
}
