% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.find.transmission.networks.from.linked.pairs}
\alias{phsc.find.transmission.networks.from.linked.pairs}
\title{Find phylogenetic transmission networks}
\usage{
phsc.find.transmission.networks.from.linked.pairs(rtp, rplkl,
  conf.cut = conf.cut, neff.cut = neff.cut, verbose = TRUE)
}
\arguments{
\item{rtp}{Pairs of individuals between whom linkage is not excluded, stored as data.table.}

\item{rplkl}{Summary of phylogenetic relationship counts for each pair, stored as data.table.}

\item{conf.cut}{Threshold on the proportion of deep-sequence phylogenies with distant/disconnected subgraphs above which pairs are considered phylogenetically unlinked. Default: 0.6}

\item{neff.cut}{Threshold on the minimum number of deep-sequence phylogenies with sufficient reads from two individuals to make any phylogenetic inferences. Default: 3.}

\item{verbose}{Flag to switch on/off verbose mode. Default: TRUE.}
}
\value{
list of two R objects 'transmission.networks', 'most.likely.transmission.chains'. See description.
}
\description{
This function reconstructs phylogenetic transmission networks from pairs of individuals between whom linkage is not excluded. 
Two R objects are generated: 
	  'transmission.networks' is a data.table describing transmission networks, with information of phylogenetic support for transmission in direction 12, direction 21, linkage with direction unclear, and no linkage. 
	  'most.likely.transmission.chains' is a data.table describing the most likely transmission chain for each transmission network.
}
