% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.get.assignments.by.window.for.couple}
\alias{phsc.get.assignments.by.window.for.couple}
\title{Get transmission assignments per window}
\usage{
phsc.get.assignments.by.window.for.couple(id1, id2, infiles)
}
\arguments{
\item{id1}{Identifier of the first individual. This is the entry for that individual in the phyloscanner .bam file.}

\item{id2}{Identifier of the second individual.}

\item{infiles}{One or more file names of transmission stats rda files. These end typically with '_trmStatsPerWindow.rda'.}
}
\value{
data.table with columns ID1 ID2 W_FROM (window start) W_TO (window end), and further columns for each infile that give the transmission assignments per window.
}
\description{
This function extracts all transmission assignments per window for two individuals.
}
\keyword{internal}
