% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.get.basic.pairwise.relationships}
\alias{phsc.get.basic.pairwise.relationships}
\title{Calculate basic pairwise relationships}
\usage{
phsc.get.basic.pairwise.relationships(df, trmw.close.brl, trmw.distant.brl,
  verbose = TRUE)
}
\arguments{
\item{df}{data.table to which basic pairwise relationships will be added. Must contain columns with name 'ADJACENT','TYPE_RAW','PATRISTIC_DISTANCE','PATHS_12','PATHS_21'.}

\item{trmw.close.brl}{Maximum patristic distance between any two read trees from both individuals in a window to classify the individuals as phylogenetically close.}

\item{trmw.distant.brl}{Minimum patristic distance between any two read trees from both individuals in a window to classify the individuals as phylogenetically distant.}
}
\value{
input data.table with new relationship column TYPE_BASIC.
}
\description{
This function calculates basic pairwise relationships of two individuals in any window. Several different relationship groups can be calculated, for example just using pairwise distance, or using both pairwise distance and topology to define likely pairs.
}
\keyword{internal}
