% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.get.most.likely.transmission.chains}
\alias{phsc.get.most.likely.transmission.chains}
\title{Reconstruct most likely transmission chains}
\usage{
phsc.get.most.likely.transmission.chains(rtnn, verbose = 0,
  method = "Edmonds")
}
\arguments{
\item{rtn}{data.table with network scores for all individuals that could form a network. Must contain columns 'ID1','ID2','IDCLU','GROUP','TYPE','POSTERIOR_SCORE','KEFF'.}
}
\value{
new data.table with added columns LINK_12 LINK_21 (either 1 or 0), and MX_PROB_12 MX_PROB_21 (associated posterior probabilities)
}
\description{
This function reconstructs most likely transmission chains 
from the scores associated with directed and undirected edges.
}
\keyword{internal}
