% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.get.pairwise.relationships.likelihoods}
\alias{phsc.get.pairwise.relationships.likelihoods}
\title{Wrapper script to calculate pairwise relationships and likelihoods}
\usage{
phsc.get.pairwise.relationships.likelihoods(dwin, trmw.min.reads, trmw.min.tips,
  trmw.close.brl, trmw.distant.brl, prior.keff, prior.neff,
  prior.calibrated.prob, relationship.types, prior.keff.dir = prior.keff,
  prior.neff.dir = prior.neff, verbose = TRUE)
}
