% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.plot.default.colours.for.relationtypes}
\alias{phsc.plot.default.colours.for.relationtypes}
\title{Default colours for relationship types}
\usage{
phsc.plot.default.colours.for.relationtypes()
}
\value{
Named list with names corresponding to RELATIONSHIP GROUPS. For each relationship group, the list contains a named vector. Entries in this vector are colours, and names are relationship types.
}
\description{
This function returns default colours for each relationship type in one of the defined relationship groups.
}
