% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.plotting.R
\name{phsc.plot.phy.selected.individuals}
\alias{phsc.plot.phy.selected.individuals}
\title{Plot short read phylogenies and highlight individuals}
\usage{
phsc.plot.phy.selected.individuals(phs, dfs, ids,
  plot.cols = rainbow(length(ids)), plot.file = NA,
  group.redo = FALSE, drop.less.than.n.ids = 2,
  drop.blacklisted = FALSE, pdf.h = 50, pdf.rw = 10,
  pdf.ntrees = 20, pdf.title.size = 40)
}
\arguments{
\item{phs}{List of trees in ape format}

\item{dfs}{data.table with mandatory column 'IDX' and optional column 'TITLE'. IDX is the index of all phylogenies in 'phs' that are to be plotted. TITLE is a title for each sub-plot, for example specifying a window.}

\item{ids}{Vector of regular expressions that identify individuals to be highlighted in colour.}

\item{plot.cols}{Vector of colours for each individual}

\item{plot.file}{If not missing, the phylogenies will be printed to file.}

\item{group.redo}{Logical, indicating if the colour groups should be recalculated from ids.}

\item{drop.blacklisted}{Logical, indicating if all blacklisted taxa should be dropped prior to plotting.}

\item{pdf.h}{Height of the pdf file in inches.}

\item{pdf.rw}{Relative width of the pdf file, internally multiplied by the number of phylogenies to give the total width in inches.}

\item{pdf.ntrees}{Number of trees per pdf.}

\item{pdf.title.size}{Size of pdf title in inches.}
}
\value{
List of ggtree objects, ready for printing.
}
\description{
This function plots short read phylogenies and highlights the clades of two individuals in red and blue.
}
