% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.plotting.R
\name{phsc.plot.phyloscan}
\alias{phsc.plot.phyloscan}
\title{Plot phyloscan}
\usage{
phsc.plot.phyloscan(rpw2, id.cols = c("ID1", "ID2"), ylim = NULL,
  cols.typet = NULL)
}
\arguments{
\item{rpw2}{data.table containing the basic phyloscanner statistics for each genomic window.}

\item{id.cols}{name of columns in rpw2 that identify the two individuals}

\item{ylim}{limits of y-axis.}

\item{cols.typet}{colour for each phylogenetic relationship type}
}
\value{
ggplot object
}
\description{
This function generates scan plots that summarize reconstructed viral phylogenetic relationships of two individuals.
Several pairs of individuals can be processed simultaneously. For each pair of individuals, the scan plot shows the 
phylogenetic distance on the y-axis and topological relationship in colours between subgraphs from both individuals in each 
deep-sequence phylogeny across the genome. The genomic position on the x-axis indicates the start of each 250bp read alignment.
}
