% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.plot.selected.pairs}
\alias{phsc.plot.selected.pairs}
\title{Plot all short read phylogenies including two individuals}
\usage{
phsc.plot.selected.pairs(phs, dfs, id1, id2, plot.file = NA, pdf.h = 50,
  pdf.rw = 10)
}
\arguments{
\item{phs}{List of trees in ape format}

\item{dfs}{data.table with mandatory column 'IDX' and optional column 'TITLE'. IDX is the index of all phylogenies in 'phs' that are to be plotted. TITLE is a title for each sub-plot, for example specifying a window.}

\item{id1}{Regular expression that identifies the first individual, to be plotted in red.}

\item{id2}{Regular expression that identifies the first individual, to be plotted in blue.}

\item{plot.file}{If not missing, the phylogenies will be printed to file.}

\item{pdf.h}{Height of the pdf file in inches.}

\item{pdf.rw}{Relative width of the pdf file, internally multiplied by the number of phylogenies to give the total width in inches.}
}
\value{
List of ggtree objects, ready for printing.
}
\description{
This function plots short read phylogenies and highlights the clades of two individuals in red and blue.
}

