% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.plotting.R
\name{phsc.plot.transmission.network}
\alias{phsc.plot.transmission.network}
\title{Plot transmission network}
\usage{
phsc.plot.transmission.network(df, di, point.size = 10,
  point.size.couple = point.size * 1.4, edge.gap = 0.04, edge.size = 0.4,
  curvature = -0.2, arrow = arrow(length = unit(0.04, "npc"), type =
  "open"), curv.shift = 0.08, label.size = 3, node.label = "ID",
  node.shape = NA_character_, node.fill = NA_character_,
  node.shape.values = NA_integer_, node.fill.values = NA_character_,
  threshold.linked = NA_real_)
}
\arguments{
\item{df}{data.table with the following columns 'IDCLU','ID1', 'ID2', 'TYPE','KEFF','LKL_MAX','POSTERIOR_SCORE'}

\item{di}{data.table with meta-data to customize the plot with columns ID, node.shape, node.label, node.fill}

\item{point.size}{size of the individual points}

\item{edge.gap}{value to adjust start / end points of edges}

\item{edge.size}{multiplier by which the size of edges is shrunk/magnified}

\item{curvature}{curvature of directed edges}

\item{arrow}{type of arrow to be plotted}

\item{curv.shift}{offset to place the label for directed edges}

\item{label.size}{size of label}

\item{node.label}{Text displayed on top of each node}

\item{node.shape}{column name in di by which the shape of each node is drawn}

\item{node.fill}{column name in di by which each node is coloured}

\item{node.shape.values}{named vector associating shapes to the values in the node.shape column}

\item{node.fill.values}{named vector associating colours to the values in the node.fill column}

\item{threshold.linked}{treshold value between 0 and 1. Edges with weight above this treshold are shown in black.}

\item{point.sizec.couple}{size of the outer ring around individuals in couples}
}
\value{
ggplot object
}
\description{
This function plots a phylogenetic transmission network, showing three types of edges: 
two directed edges respectively in the 1->2 and 2->1 direction, and an undirected edge that represents phylogenetic support of close and adjacent individuals without evidence into the direction transmission.
}
