% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.plot.windowsummaries.for.pairs}
\alias{phsc.plot.windowsummaries.for.pairs}
\title{Plot window summaries for pairs of individuals}
\usage{
phsc.plot.windowsummaries.for.pairs(plot.select, rpw2, rplkl2, plot.file,
  cols = NULL, group = NA, widths = unit(c(4, 6), "null"),
  heights = unit(c(2, 3.5, 4, 5), "null"), height = 9)
}
\arguments{
\item{plot.select}{Select pairs for plotting. This is specified as a data.table with columns PTY_RUN, MALE_SANGER_ID, FEMALE_SANGER_ID, LABEL. Each pair has assigned a label that is used as title.}

\item{rpw2}{Window assignments for pairs of individuals. This is specified as a data.table with columns PTY_RUN, MALE_SANGER_ID, FEMALE_SANGER_ID, W_FROM, ID_R_MAX, ID_R_MIN, TYPE.}

\item{rplkl2}{Posterior probability assignments for pairs of individuals. This is specified as a data.table with columns PTY_RUN, MALE_SANGER_ID, FEMALE_SANGER_ID, TYPE, KEFF, POSTERIOR_ALPHA, POSTERIOR_BETA}

\item{plot.file}{Name of output file}

\item{cols}{Colours for relationship types. This is specified as a named vector, which can be null.}

\item{group}{Relationship group name. This is used to define default colours if cols==NULL.}

\item{heights}{Heights of subfigures in grid.layout. This is overwritten if group is specified.}

\item{height}{Total height of pdf in inches. This is overwritten if group is specified.}

\item{widhts}{Widths of subfigures in grid.layout. This is overwritten if group is specified.}
}
\value{
Plots to file.
}
\description{
This function plots the window summaries for each pair of individuals.
}
\keyword{internal}
