% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.read.likelytransmissions}
\alias{phsc.read.likelytransmissions}
\title{Read likely transmissions summary files into a data.table}
\usage{
phsc.read.likelytransmissions(prefix.infiles, prefix.run = "ptyr",
  regexpr.lklsu = "trmStats.csv$", save.file = NA, resume = FALSE,
  zip = FALSE)
}
\arguments{
\item{prefix.infiles}{Full path name identifying likely transmissions summary files}

\item{prefix.run}{Character string to identify separate phyloscanner runs. After the prefix, an integer number is expected. For example, if prefix.run='ptyr', then files are expected to start like 'ptyr123_'.}

\item{regexpr.lklsu}{Regular expression that identifies likely transmissions summary files in the directory.}

\item{save.file}{If not missing, function output (a data.table) will be stored to 'save.file', input files will be zipped, and input files will be deleted.}

\item{resume}{If TRUE and save.file is not missing, the function loads and returns trees stored in save.file.}
}
\value{
Data table with columns PAIR_ID, ID1, ID2, TYPE, WIN_OF_TYPE, PTY_RUN, WIN_TOTAL, SCORE
}
\description{
This function reads likely transmissions summary files from the phyloscanner toolkit.
}
