% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.read.processed.phyloscanner.output.in.directory}
\alias{phsc.read.processed.phyloscanner.output.in.directory}
\title{Read processed phyloscanner output}
\usage{
phsc.read.processed.phyloscanner.output.in.directory(in.dir, save.file.base,
  resume = FALSE, zip = FALSE)
}
\arguments{
\item{in.dir}{Full path name to processed phyloscanner output}

\item{save.file.base}{Output will be stored to files that start with 'save.file.base'.}

\item{resume}{If TRUE, the function does not process existing rda files.}

\item{zip}{If TRUE, the function zips processed phyloscanner output, and then deletes the zipped, processed phyloscanner output files.}
}
\value{
Nothing, rda objects written to file.
}
\description{
This function creates R data.tables from processed phyloscanner output, for further data analysis in R.
}

