% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.scripting.R
\name{raxml.cmd}
\alias{raxml.cmd}
\title{Generate bash commands to generate phylogenies with RAxML}
\usage{
raxml.cmd(infile.fasta, outfile = paste(infile.fasta, ".newick", sep = ""),
  pr = "raxmlHPC-SSE3", pr.args = "-m GTRCAT --HKY85 -p 42")
}
\arguments{
\item{infile.fasta}{Full path name to input fasta file.}

\item{outfile}{Full path name to output file. All RAxML output other than the best tree will be zipped and returned as 'outfile.zip'.}

\item{pr}{Full path name to RAXML program.}

\item{pr.args}{RAXML arguments.}
}
\value{
Character string
}
\description{
This function generates bash commands to generate one phylogeny with RAxML, that can be called via 'system' in R, or written to file to run on a UNIX system.
}
