% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.cmd.phyloscanner.one}
\alias{phsc.cmd.phyloscanner.one}
\title{Generate bash command for a single phyloscanner run}
\usage{
phsc.cmd.phyloscanner.one(file.bam, file.ref, window.coord = integer(0),
  window.automatic = "", prog = PROG.PTY, prog.raxml = "raxmlHPC-AVX",
  prog.mafft = "mafft", merge.threshold = 1, min.read.count = 2,
  quality.trim.ends = 30, min.internal.quality = 2,
  merge.paired.reads = TRUE, num.bootstraps = 1,
  all.bootstrap.trees = TRUE, no.trees = FALSE, keep.overhangs = FALSE,
  dont.check.duplicates = FALSE, file.alignments = NA, root = NA,
  align.pairwise.to = NA, out.dir = ".")
}
\description{
This function generates bash commands for a single phyloscanner run, that can be called via 'system' in R, or written to file to run on a UNIX system.
}

