% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.R
\name{phsc.cmd.process.phyloscanner.output.in.directory}
\alias{phsc.cmd.process.phyloscanner.output.in.directory}
\title{Generate bash commands to process phyloscanner output}
\usage{
phsc.cmd.process.phyloscanner.output.in.directory(ptyf, pty.args)
}
\arguments{
\item{ptyf}{data.table with columns PTY_RUN and FILE_PTY_RUN. PTY_RUN is an id given to a single phyloscanner run. FILE_PTY_RUN is a file base name that identifies phyloscanner output from that phyloscanner run.}

\item{pty.args}{phyloscanner input arguments that were used to obtain the phyloscanner output listed in 'ptyf'.}
}
\value{
data.table with one column CMD that contains the bash commands.
}
\description{
This function generates bash commands that combine the various Rscript in the phyloscanner toolkit
}

