% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloscan.fun.private.R
\name{phsc.read.processed.phyloscanner.output.in.directory}
\alias{phsc.read.processed.phyloscanner.output.in.directory}
\title{Read processed phyloscanner output}
\usage{
phsc.read.processed.phyloscanner.output.in.directory(
  prefix.infiles,
  save.file.base,
  read.likelytransmissions = TRUE,
  read.trees = TRUE,
  read.subtrees = TRUE,
  resume = FALSE,
  zip = FALSE
)
}
\arguments{
\item{prefix.infiles}{Full path name to processed phyloscanner output}

\item{save.file.base}{Output will be stored to files that start with 'save.file.base'.}

\item{read.likelytransmissions}{If TRUE, read and process likely transmissions}

\item{read.trees}{If TRUE, read and process trees}

\item{read.subtrees}{If TRUE, read and process subtree files}

\item{resume}{If TRUE, the function does not process existing rda files.}

\item{zip}{If TRUE, the function zips processed phyloscanner output, and then deletes the zipped, processed phyloscanner output files.}
}
\value{
Nothing, rda objects written to file.
}
\description{
This function creates R data.tables from processed phyloscanner output, for further data analysis in R.
}
\keyword{internal}
